// src/features/ResumeColors/ResumeColors.interfaces.ts
export interface palitra {
  color: string | undefined
  mainText: string | undefined
  fon: string | undefined
  backgroundColor: string | undefined
  backgroundMainColor: string | undefined
  dateColor: string | undefined
  title: string | undefined
  borderSidebar: string | undefined
  borderMain: string | undefined
}

export interface Color {
  showColor: string
  color: string
  mainText: string
  backgroundColor: string
  dateColor: string
  backgroundMainColor: string
  title: string
  borderSidebar: string
  borderMain: string
  fon: string
}